Use promise in getting the schema by schema provider (#92)
Since ZIO effects work differently from Futures, a bug was introduced:

We created a SchemaProvider instance for each folder to be able to get the schema for each row. Previously, we executed the Future that completed with the schema result. When it was converted to a ZIO effect, it led to querying the schema for every row.

This PR changes that. Instead of running the effect on every row, we are running a ZIO.Promise for each row, which works the same way as scala.concurrent.Future.

See ZIO.Promise documentation here: https://zio.dev/reference/concurrency/promise/